fix: 결제 대기 상태에서 예약을 삭제하면 payment 정보 삭제하지 않도록 수정

+++ src/main/java/roomescape/service/reservation/ReservationService.java
@@ -165,7 +165,9 @@ public class ReservationService {
         Reservation reservation = reservationRepository.getReservationById(reservationId);
         validateReservationMember(reservation, member);
 
-        paymentService.deleteReservationPayment(reservation);
+        if (!reservation.isPaymentWaitingStatus()) {
+            paymentService.deleteReservationPayment(reservation);
+        }
         reservationRepository.delete(reservation);
     }
